Simplify libhover devhelp preferences page
* Remove unnecessary LabelFieldEditor
* Remove duplicate generation job
* Move generation job into it's own class from DevHelpStartup and re-use
  it from the preferences page to avoid duplication
